Divided con_lib into seperate failes
